LayoutTests/accessibility/render-counter-text.html

<!DOCTYPE HTML>
<html>
<body>
<script src="../fast/js/resources/js-test-pre.js"></script>

<style>
ol li::before { content: counter(l1) ":"; counter-increment: l1;}
ol li { list-style-type: decimial; display: block; }
</style>

<div id="container" style="counter-reset:l1;">
<ol style="list-style-type:decimal">
<li id="line1">line1
<li id="line2">line2
<li>line3
</ol>
</div>

<div id="description"></div>
<div id="console"></div>

<script>

description("This tests that the text for RenderCounter is exposed correctly.")

if (window.testRunner && window.accessibilityController) {
    
    var line1 = accessibilityController.accessibleElementById("line1");
    shouldBe("line1.childAtIndex(0).stringValue", "'AXValue: 1'");
    shouldBe("line1.childAtIndex(1).stringValue", "'AXValue: :'");
    shouldBe("line1.childAtIndex(2).stringValue", "'AXValue: line1'");

    var line2 = accessibilityController.accessibleElementById("line2");
    shouldBe("line2.childAtIndex(0).stringValue", "'AXValue: 2'");
    shouldBe("line2.childAtIndex(1).stringValue", "'AXValue: :'");
    shouldBe("line2.childAtIndex(2).stringValue", "'AXValue: line2'");

    document.getElementById("container").style.display = "none";
}

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
